requirements.txt
# Core dependencies
streamlit>=1.28.0

# API client for Groq
groq>=0.4.0

# Embeddings and vector search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# PDF processing
pypdf>=4.0.0

# LangChain components (minimal)
langchain>=0.1.0
langchain-community>=0.0.20

# Data processing
numpy>=1.24.0

# Environment variables
python-dotenv>=1.0.0
